Address Jiayin's comments
