+++ src/common/eventType.ts
@@ -56,7 +56,6 @@ export enum EventType {
 
     // Language Server
     StartLanguageServerCompleted = 'StartLanguageServerCompleted',
-    CredentialExpiredDuringLanguageServerRunning = 'CredentialExpiredDuringLanguageServerRunning',
 
     // Test only
     RefreshCredential = 'RefreshCredential',
