+++ src/observers/infoMessageObserver.ts
@@ -33,9 +33,6 @@ export class InfoMessageObserver {
             case EventType.ExtensionActivated:
                 this.handleExtensionActivated();
                 break;
-            case EventType.CredentialExpiredDuringLanguageServerRunning:
-                this.handleCredentialExpiredWhenLanguageServerRunning();
-                break;
         }
     }
 
@@ -89,11 +86,5 @@ export class InfoMessageObserver {
                     }
                 ));
         }
-    }
-
-    private handleCredentialExpiredWhenLanguageServerRunning() {
-        const action = new MessageAction('Sign in', 'docs.signIn');
-        const message = 'Credential Expired, please sign in again.';
-        this.showInfoMessage(message, action);
     }
 }
